Verify Admin

+++ routes/levelRouter.js
@@ -1,8 +1,8 @@
 const express = require("express");
 const Level = require("../models/levelSchema");
-const levelRouter = express.Router();
 const authenticate = require("../authenticate");
 
+const levelRouter = express.Router();
 
 levelRouter
   .route("/")
@@ -19,7 +19,7 @@ levelRouter
   .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Level.create(req.body)
       .then((level) => {
-        console.log("level Created ", level);
+        console.log("Level Created ", level);
         res.statusCode = 200;
         res.setHeader("Content-Type", "application/json");
         res.json(level);
@@ -54,7 +54,9 @@ levelRouter
   })
   .post(authenticate.verifyUser, (req, res) => {
     res.statusCode = 403;
-    res.end(`POST operation not supported on /levels/${req.params.levelId}`);
+    res.end(
+      `POST operation not supported on /levels/${req.params.levelId}`
+    );
   })
   .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Level.findByIdAndUpdate(
@@ -71,15 +73,19 @@ levelRouter
       })
       .catch((err) => next(err));
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Level.findByIdAndDelete(req.params.levelId)
-      .then((response) => {
-        res.statusCode = 200;
-        res.setHeader("Content-Type", "application/json");
-        res.json(response);
-      })
-      .catch((err) => next(err));
-  });
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Level.findByIdAndDelete(req.params.levelId)
+        .then((response) => {
+          res.statusCode = 200;
+          res.setHeader("Content-Type", "application/json");
+          res.json(response);
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 levelRouter
   .route("/:levelId/feedbacks")
@@ -127,29 +133,33 @@ levelRouter
       `PUT operation not supported on /levels/${req.params.levelId}/feedbacks`
     );
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Level.findById(req.params.levelId)
-      .then((level) => {
-        if (level) {
-          for (let i = level.feedbacks.length - 1; i >= 0; i--) {
-            level.feedbacks.id(level.feedbacks[i]._id).deleteOne();
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Level.findById(req.params.levelId)
+        .then((level) => {
+          if (level) {
+            for (let i = level.feedbacks.length - 1; i >= 0; i--) {
+              level.feedbacks.id(level.feedbacks[i]._id).deleteOne();
+            }
+            level
+              .save()
+              .then((level) => {
+                res.statusCode = 200;
+                res.setHeader("Content-Type", "application/json");
+                res.json(level);
+              })
+              .catch((err) => next(err));
+          } else {
+            err = new Error(`Level ${req.params.levelId} not found`);
+            err.status = 404;
+            return next(err);
           }
-          level
-            .save()
-            .then((level) => {
-              res.statusCode = 200;
-              res.setHeader("Content-Type", "application/json");
-              res.json(level);
-            })
-            .catch((err) => next(err));
-        } else {
-          err = new Error(`Level ${req.params.levelId} not found`);
-          err.status = 404;
-          return next(err);
-        }
-      })
-      .catch((err) => next(err));
-  });
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 levelRouter
   .route("/:levelId/feedbacks/:feedbackId")
@@ -179,24 +189,36 @@ levelRouter
       `POST operation not supported on /levels/${req.params.levelId}/feedbacks/${req.params.feedbackId}`
     );
   })
+
   .put(authenticate.verifyUser, (req, res, next) => {
     Level.findById(req.params.levelId)
       .then((level) => {
         if (level && level.feedbacks.id(req.params.feedbackId)) {
-          if (req.body.rating) {
-            level.feedbacks.id(req.params.feedbackId).rating = req.body.rating;
+          if (
+            level.feedbacks
+              .id(req.params.feedbackId)
+              .author._id.equals(req.user._id)
+          ) {
+            if (req.body.rating) {
+              level.feedbacks.id(req.params.feedbackId).rating =
+                req.body.rating;
+            }
+            if (req.body.text) {
+              level.feedbacks.id(req.params.feedbackId).text = req.body.text;
+            }
+            level
+              .save()
+              .then((level) => {
+                res.statusCode = 200;
+                res.setHeader("Content-Type", "application/json");
+                res.json(level);
+              })
+              .catch((err) => next(err));
+          } else {
+            err = new Error(`Not authorized`);
+            err.status = 403;
+            return next(err);
           }
-          if (req.body.text) {
-            level.feedbacks.id(req.params.feedbackId).text = req.body.text;
-          }
-          level
-            .save()
-            .then((level) => {
-              res.statusCode = 200;
-              res.setHeader("Content-Type", "application/json");
-              res.json(level);
-            })
-            .catch((err) => next(err));
         } else if (!level) {
           err = new Error(`Level ${req.params.levelId} not found`);
           err.status = 404;
@@ -213,26 +235,38 @@ levelRouter
     Level.findById(req.params.levelId)
       .then((level) => {
         if (level && level.feedbacks.id(req.params.feedbackId)) {
-          level.feedbacks.id(req.params.feedbackId).deleteOne();
-          level
-            .save()
-            .then((level) => {
-              res.statusCode = 200;
-              res.setHeader("Content-Type", "application/json");
-              res.json(level);
-            })
-            .catch((err) => next(err));
+          if (
+            level.feedbacks
+              .id(req.params.feedbackId)
+              .author._id.equals(req.user._id)
+          ) {
+            level.feedbacks.id(req.params.feedbackId).deleteOne();
+            level
+              .save()
+              .then((level) => {
+                res.statusCode = 200;
+                res.setHeader("Content-Type", "application/json");
+                res.json(level);
+              })
+              .catch((err) => next(err));
+          } else {
+            err = new Error(`Not authorized`);
+            err.status = 403;
+            return next(err);
+          }
         } else if (!level) {
           err = new Error(`Level ${req.params.levelId} not found`);
           err.status = 404;
           return next(err);
         } else {
-          err = new Error(`feedback ${req.params.feedbackId} not found`);
+          err = new Error(`Feedback ${req.params.feedbackId} not found`);
           err.status = 404;
           return next(err);
         }
       })
       .catch((err) => next(err));
   });
+
+module.exports = levelRouter;
 
 module.exports = levelRouter;
